# apps/server/openapi.yaml
openapi: 3.0.3
info:
  title: Electronic Lab Notebook API
  description: |
    REST API for the Electronic Lab Notebook (ELN) application.
    
    ## Authentication
    
    All endpoints require authentication via API key or user session:
    - **API Key**: Include `x-api-key` header with your API key
    - **Session**: Include `x-user-id` header with your user ID
    
    ## Rate Limiting
    
    API calls are rate limited to 1000 requests per hour per API key.
    
    ## Pagination
    
    List endpoints support pagination via `limit` and `offset` query parameters.
    Default limit is 50, maximum is 200.
    
    ## Data Export
    
    All list endpoints support data export via `format` query parameter:
    - `json` (default)
    - `csv`
    - `xlsx`
  version: 1.0.0
  contact:
    name: ELN Support
    email: support@eln.local
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000
    description: Local development server
  - url: https://eln.example.com/api
    description: Production server
...
paths:
...
  /export/experiments:
    get:
      summary: Export experiments data
      tags: [export]
      parameters:
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [json, csv, xlsx, pdf, zip]
        - name: ids
          in: query
          schema:
            type: array
            items:
              type: string
              format: uuid
          description: Specific experiment IDs to export (all if omitted)
        - name: includeAttachments
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Exported data
          content:
            application/json:
              schema:
                type: object
            text/csv:
              schema:
                type: string
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
            application/zip:
              schema:
                type: string
                format: binary

  # ==================== API KEYS ====================
  /api-keys:
    get:
      summary: List user's API keys
      tags: [users]
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/APIKey'
    post:
      summary: Create new API key
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, permissions]
              properties:
                name:
                  type: string
                permissions:
                  type: array
                  items:
                    $ref: '#/components/schemas/APIPermission'
                expiresAt:
                  type: string
                  format: date-time
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                    description: The API key (only shown once)
                  apiKey:
                    $ref: '#/components/schemas/APIKey'
  /api-keys/{id}:
    delete:
      summary: Revoke API key
      tags: [users]
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '204':
          description: API key revoked

  # ==================== SYNC ====================
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
    userId:
      type: apiKey
      in: header
      name: x-user-id
  parameters:
    idParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    APIPermission:
      type: string
      enum: [read:experiments, write:experiments, read:methods, write:methods, read:inventory, write:inventory, read:attachments, write:attachments, read:users, admin]
...
    APIKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        keyPrefix:
          type: string
          description: First 8 characters of the key for identification
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/APIPermission'
        expiresAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
security:
  - apiKey: []
  - userId: []